Add Code to Add new Entry *Added demo entries to db for testing

+++ src/main/java/com/courier/management/dto/EntryDTO.java
@@ -22,7 +22,7 @@ public class EntryDTO {
 
     private Long customerId;
 
-    private Long clientId;
+    private Integer clientId;
 
     private Long invoiceId; // nullable
